refactor: 커피챗 서비스 공통 메서드 CoffeeChatCommon으로 분리
- CoffeeChatApplicationService, CoffeeChatInfoService, CoffeeChatReceivedService에서
  중복으로 사용되던 공통 로직을 CoffeeChatCommon 클래스로 분리해 재사용성 및 유지보수성 향상
