+++ boottalk/src/main/java/com/icandoit/boottalk/coffeeChat/service/CoffeeChatInfoService.java
@@ -1,5 +1,7 @@
 package com.icandoit.boottalk.coffeeChat.service;
 
+import org.springframework.stereotype.Service;
+
 import com.icandoit.boottalk.coffeeChat.dto.CoffeeChatInfoRequestDto;
 import com.icandoit.boottalk.coffeeChat.dto.CoffeeChatInfoResponseDto;
 import com.icandoit.boottalk.coffeeChat.entity.CoffeeChatInfo;
@@ -7,23 +9,21 @@ import com.icandoit.boottalk.coffeeChat.repository.CoffeeChatInfoRepository;
 import com.icandoit.boottalk.libs.exception.CustomException;
 import com.icandoit.boottalk.libs.exception.ErrorCode;
 import com.icandoit.boottalk.user.domain.entity.User;
-import com.icandoit.boottalk.user.domain.repository.UserRepository;
+
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
 public class CoffeeChatInfoService {
 
-    private final UserRepository userRepository;
     private final CoffeeChatInfoRepository coffeeChatInfoRepository;
+
+    private final CoffeeChatCommonService coffeeChatCommonService;
 
     public CoffeeChatInfoResponseDto createCoffeeChatInfo(Long userId,
         CoffeeChatInfoRequestDto requestDto) {
 
-        User user = userRepository.findById(userId)
-            .orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+        User user = coffeeChatCommonService.getUser(userId);
 
         if (coffeeChatInfoRepository.existsByMentor_UserId(userId)) {
             throw new CustomException(ErrorCode.COFFEE_CHAT_ALREADY_EXISTS);
@@ -42,7 +42,7 @@ public class CoffeeChatInfoService {
 
     public CoffeeChatInfoResponseDto getMyCoffeeChatInfo(Long userId) {
 
-        CoffeeChatInfo coffeeChatInfo = getCoffeeChatInfoByUserId(userId);
+        CoffeeChatInfo coffeeChatInfo = coffeeChatCommonService.getCoffeeChatInfoByUserId(userId);
         return CoffeeChatInfoResponseDto.from(coffeeChatInfo);
     }
 
@@ -56,14 +56,14 @@ public class CoffeeChatInfoService {
     public CoffeeChatInfoResponseDto updateMyCoffeeChatInfo(
         Long userId, CoffeeChatInfoRequestDto requestDto) {
 
-        CoffeeChatInfo coffeeChatInfo = getCoffeeChatInfoByUserId(userId);
+        CoffeeChatInfo coffeeChatInfo = coffeeChatCommonService.getCoffeeChatInfoByUserId(userId);
 
         coffeeChatInfo.update(requestDto);
         return CoffeeChatInfoResponseDto.from(coffeeChatInfo);
     }
 
     public void deleteMyCoffeeChatInfo(Long userId) {
-        CoffeeChatInfo coffeeChatInfo = getCoffeeChatInfoByUserId(userId);
+        CoffeeChatInfo coffeeChatInfo = coffeeChatCommonService.getCoffeeChatInfoByUserId(userId);
 
         // 멘토링 활동 금지 당한 커피챗 정보는 삭제 불가
         if (coffeeChatInfo.isMentoringBanned()) {
@@ -73,8 +73,4 @@ public class CoffeeChatInfoService {
         coffeeChatInfoRepository.delete(coffeeChatInfo);
     }
 
-    protected CoffeeChatInfo getCoffeeChatInfoByUserId(Long userId) {
-        return coffeeChatInfoRepository.findByMentor_UserId(userId)
-            .orElseThrow(() -> new CustomException(ErrorCode.USER_COFFEE_CHAT_NOT_FOUND));
-    }
 }
